Possible fix for segmentation fault in FMA code path
The FMA code caused a segmentation fault when calling _mm256_mul_ps() in
CondLikeDown_NUC4_FMA().  This was likely caused by using FMA SIMD
instructions on misaligned data.

Upon debugging this, I noticed that AlignedMalloc() was being called
with an alignment of 16 rather than 32 for some of the data.  This was
due to two if-statements only testing for AVX and not also for FMA.

The patch fixes the two if-statements and the resulting binary no longer
exhibit the segmentation fault.

I noted that another if-statement already had this extra test for FMA
included.

This is primarily relating to PR #74 and possibly also to PR #65.