Make sure that the correct beagle-lib tag is checked out
It was pointed out in issue #254 that the most recent beagle-lib
releases are using CMake and therefore have a somewhat different
building procedure compared to before.  This PR "solves" this by asking
our users to use the "v3.1.2" tag of the project, which we know works
and that we have tested MrBayes with in the past.

Note that this only affects uses that actually need to build beagle-lib
locally using instructions that we know works.  Any user is obviously
still free to use a later release of beagle-lib, and to build it in any
way they want.  We just don't know if that will work or not.

This resolves #254.